Added a way to add the name of the card dinamically into the desc.
If the description contains $THIS_NAME, it will be replaced with the
name of the card it goes on. This is useful for development, because the
name of the card might change slightly, and using this feature will
prevent the description from be out of date.

Other accepted strings are:
- $THIS_CARD
- $this_name
- $this_card