v2.5.7
Added support for patterns in the .ods file. You can use a Java regex
pattern for the name (not the folder) and it will look for all matches
of that pattern, then pick the most recently modified one. Good if you
download new versions of the .ods file after editing it only or
something.